sonic-architect/requirements.txt
altair==6.0.0
attrs==26.1.0
audioop-lts==0.2.2
audioread==3.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.3.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
msgpack==1.1.2
narwhals==2.19.0
numba==0.65.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.4
pooch==1.9.0
protobuf==7.34.1
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.2
pydub==0.25.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.3
soundfile==0.13.1
soxr==1.0.0
spotipy==2.26.0
SQLAlchemy==2.0.49
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
yt-dlp==2026.3.17
